完善用户删除，修改，新增，功能。增加处理Xss注入

+++ xh-ssm/system-admin/src/main/java/com/xh/annotation/Repeat.java
@@ -9,6 +9,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Repeat {
 
-    int value() default 10;
+    long value() default 10;
 
 }
